feat: rewards refactor
