+++ assistant/rewards/classes/StakeEvent.py
