+++ scripts/status/retroactive_xsushi_distribution.py
@@ -1,17 +1,18 @@
+import json
+
 from brownie import *
 from tqdm import tqdm
 from config.badger_config import badger_config
 from rich.console import Console
 from scripts.systems.badger_system import connect_badger
-import json
-from assistant.rewards.rewards_assistant import calc_sushi_rewards,process_cumulative_rewards,fetch_current_rewards_tree
-from assistant.rewards.rewards_checker import test_claims
-from assistant.subgraph.client import fetch_harvest_farm_events
-from assistant.rewards.RewardsList import RewardsList
+from assistant.rewards.classes.RewardsList import RewardsList
 from config.rewards_config import rewards_config
 from brownie.network.gas.strategies import GasNowStrategy
-from assistant.rewards.merkle_tree import rewards_to_merkle_tree
-from assistant.rewards.RewardsLogger import rewardsLogger
+from assistant.rewards.classes.MerkleTree import rewards_to_merkle_tree
+from assistant.rewards.classes.RewardsLogger import rewardsLogger
+from assistant.rewards.rewards_assistant import fetch_current_rewards_tree
+from assistant.rewards.rewards_utils import process_cumulative_rewards
+from assistant.rewards.meta_rewards.sushi import calc_all_sushi_rewards
 
 gas_strategy = GasNowStrategy("fast")
 console = Console()
@@ -20,9 +21,9 @@ def main():
     test = True
     badger = connect_badger(badger_config.prod_json,load_deployer=False)
     nextCycle = badger.badgerTree.currentCycle() + 1
-    startBlock = 11537600
+    startBlock = 0
     endBlock = chain.height
-    rewards = calc_sushi_rewards(badger,startBlock,endBlock,nextCycle,retroactive=True)
+    rewards = calc_all_sushi_rewards(badger,startBlock,endBlock,nextCycle,retroactive=True)
     rewardsLogger.save("retroactive-xsushi")
     currentRewards = fetch_current_rewards_tree(badger)
